requirements.txt
django
feincms
django-mptt
south
lxml
pillow
BeautifulSoup
feedparser
django-debug-toolbar
django-simple-captcha
django-imagekit
django-appconf
#django-permission
#six
#pytz
#pytidylib
#gunicorn
docutils
pygments
sphinx
markdown
#django-sekizai
#sorl-thumbnail
